historic model and integration with account app + retrieval model

+++ retrieval/src/main/java/com/crypto_tracker/retrieval/coins/CurrentPrice.java
@@ -11,5 +11,5 @@ import lombok.NoArgsConstructor;
 @JsonIgnoreProperties
 public class CurrentPrice {
 
-    private int usd;
+    private long usd;
 }
